+++ exercises/hard/solutiont4/src/utils.rs
@@ -7,7 +7,7 @@ use crate::iso_week;
 ///
 /// # 返回值
 /// - `true` 如果是闰年，否则返回 `false`
-pub fn is_leap_year(year: i32) -> bool {
+pub fn is_leap_year(year: u32) -> bool {
     // 判断年份是否满足闰年的条件
     (year % 4 == 0 && year % 100 != 0) || (year % 400 == 0)
 }
@@ -19,7 +19,7 @@ pub fn is_leap_year(year: i32) -> bool {
 ///
 /// # 返回值
 /// - 返回一个包含每月天数的数组，若是闰年，则二月为29天
-pub fn compute_days_in_month(year: i32) -> [i32; 12] {
+pub fn compute_days_in_month(year: u32) -> [u32; 12] {
     // 定义非闰年每个月的天数
     let mut days_in_month = [31, 28, 31, 30, 31, 30, 31, 31, 30, 31, 30, 31];
     // 如果是闰年则修改2月天数为29天
@@ -40,13 +40,13 @@ pub fn compute_days_in_month(year: i32) -> [i32; 12] {
 /// - `(current_day, total_days)`
 ///   - `current_day`: 指定日期是当年的第几天
 ///   - `total_days`: 当年的总天数（365或366）
-pub fn get_year_info(year: i32, month: usize, day: i32) -> (i32, i32) {
+pub fn get_year_info(year: u32, month: u32, day: u32) -> (u32, u32) {
     // 先获取每个月的天数数组
     let days_in_month = compute_days_in_month(year);
     // 累加前几个月的天数，再加上当前月份的日期，得到当前天数
-    let current_day: i32 = days_in_month.iter().take(month - 1).sum::<i32>() + day;
+    let current_day: u32 = days_in_month.iter().take((month - 1) as usize).sum::<u32>() + day;
     // 全年总天数为各个月天数之和
-    let total_days: i32 = days_in_month.iter().sum();
+    let total_days: u32 = days_in_month.iter().sum();
     (current_day, total_days)
 }
 
@@ -59,7 +59,7 @@ pub fn get_year_info(year: i32, month: usize, day: i32) -> (i32, i32) {
 ///
 /// # 返回值
 /// - 返回该日期是当年的第几天
-pub fn day_of_year(year: i32, month: usize, day: i32) -> i32 {
+pub fn day_of_year(year: u32, month: u32, day: u32) -> u32 {
     // 获取当前天数（当年的第几天）
     get_year_info(year, month, day).0
 }
@@ -71,7 +71,7 @@ pub fn day_of_year(year: i32, month: usize, day: i32) -> i32 {
 ///
 /// # 返回值
 /// - 返回该年份的总天数（365或366）
-pub fn days_in_year(year: i32) -> i32 {
+pub fn days_in_year(year: u32) -> u32 {
     // 获取每个月的天数并求和，得出全年总天数
     compute_days_in_month(year).iter().sum()
 }
@@ -85,11 +85,11 @@ pub fn days_in_year(year: i32) -> i32 {
 ///
 /// # 返回值
 /// - `(next_year, next_month, next_day)` 表示下一天的日期
-pub fn next_day(year: i32, month: usize, day: i32) -> (i32, usize, i32) {
+pub fn next_day(year: u32, month: u32, day: u32) -> (u32, u32, u32) {
     // 获取当前年份中每个月的天数
     let days_in_month = compute_days_in_month(year);
     // 如果当前日期未达到当月最大天数，则次日只需要增加一天
-    if day < days_in_month[month - 1] {
+    if day < days_in_month[(month - 1) as usize] {
         (year, month, day + 1)
     // 如果不是12月份，则进入下个月，并将日期设为1号
     } else if month < 12 {
@@ -109,7 +109,7 @@ pub fn next_day(year: i32, month: usize, day: i32) -> (i32, usize, i32) {
 ///
 /// # 返回值
 /// - `true` 如果该日期为周六或周日，否则返回 `false`
-pub fn is_weekend(year: i32, month: usize, day: i32) -> bool {
+pub fn is_weekend(year: u32, month: u32, day: u32) -> bool {
     // 获取日期对应的星期，返回1至7对应周一到周日
     let dow = iso_week::day_of_week_sakamoto(year, month, day);
     // 判断星期是否为6（周六）或7（周日）
